Revert "server/user: eager loading of oauth accounts"
This reverts commit b8596a097cda219b6b5ee4193137e6fba2f36eb9.
